% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_acres.R
\name{calculate_acres}
\alias{calculate_acres}
\title{Calculate acres for a simple feature (sf) polygon object.}
\usage{
calculate_acres(sf_poly, geodesic = TRUE)
}
\arguments{
\item{sf_poly}{\code{sf} polygon object.}

\item{geodesic}{Optional. TRUE/FALSE. Calculate geodesic acres if input
polygon has a geographic CRS. Default is TRUE.}
}
\value{
An \code{sf} polygon object.
}
\description{
This function evaluates if the units of the input polygon are in meters
before calculating acres. If the units are are not in meters, acres are
not calculated. Then the function will evaluate if geodesic or planar
acres can be calculated. Geodesic acres are calculated if the input
polygon is has a geographic coordinate reference system (CRS; e.g.,
WGS84, EPSG:4326). If the input polygon does not have a geographic CRS,
planar acres will be calculated.
}
\examples{
\dontrun{
library("psoGIStools")

dat_sf <- read_fc(lyr_name = "feature_name",
                  dsn = file.path("T:/path/to/geodatabase"),
                  crs = "EPSG:4326") |> 
          calculate_acres()

}
}
\seealso{
\code{\link[sf:st_area]{sf::st_area()}}, \code{\link[sf:sf_use_s2]{sf::sf_use_s2()}}
}
