% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve_sf.R
\name{dissolve_sf}
\alias{dissolve_sf}
\title{Dissolve a simple feature}
\usage{
dissolve_sf(input_feature, var_names)
}
\arguments{
\item{input_feature}{Input spatial feature (\code{sf} object).}

\item{var_names}{Vector of variables names to dissolve data by.}
}
\value{
An \link[sf:sf]{sf} object.
}
\description{
This function performs a dissolve, just like ArcGIS, by using
\code{dplyr::summarise()} and grouping the summary by the variable names
provided by the use.
}
\examples{
\dontrun{
library(psoGIStools)

# Read data from Forest Service ArcGIS REST Service
fs_lands <- read_edw_lyr("EDW_SurfaceOwnership_01") |>
  dplyr::filter(ownerclassification == 'USDA FOREST SERVICE')

# Use `dissolve_sf()` to dissolve by region and unit name. 
fs_units <- dissolve_sf(input_feature = fs_lands, 
                        var_names = c("region", "nfslandunitname"))
}
}
