% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fc.R
\name{view_layers}
\alias{view_layers}
\title{View layers in a geodatabase}
\usage{
view_layers(dsn)
}
\arguments{
\item{dsn}{Path to geodatabase.}
}
\value{
A vector of layer or feature class names.
}
\description{
This function retuns a vecor of layer or feature class names in a geodatabase.
}
\examples{
\dontrun{
gdb <- file.path("data", "MBF_SppEvals.gdb")
view_layers(gdb)
}
}
