% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_sf.R
\name{clip_sf}
\alias{clip_sf}
\title{Clip simple feature to polygon}
\usage{
clip_sf(input_feature, clip_feature)
}
\arguments{
\item{input_feature}{Input spatial feature (\code{sf} object) to be clipped.}

\item{clip_feature}{Polygon feature (\code{sf}) used to clip \code{input_feature}.}
}
\value{
An \link[sf:sf]{sf} object.
}
\description{
This function clips an \code{sf} object, just like ArcGIS, by using
\code{sf::st_intersection()}. First, this function checks that the coordinate
reference system (CRS) of the input object is the same as the clipping
object. If it is not, this function transforms the clipping object to CRS
of the input object using \code{sf::st_transform()} before clipping. The
output CRS is not changed.
}
\examples{
\dontrun{
library("psoGIStools")
library("dplyr")

#-- Read spatial data into R
# Fishlake National Forest Adminstrative Boundary
fif <- read_edw_lyr("EDW_ForestSystemBoundaries_01") |>
  filter(forestname == "Fishlake National Forest")

# Recreation Sites
rec_sites <- read_edw_lyr("EDW_InfraRecreationSites_01")

#-- Clip rec sites to forest boundary
fif_rec_sties <- clip_sf(rec_sites, fif)
}
}
\seealso{
\code{\link[=read_fc]{read_fc()}}, \code{\link[sf:st_intersection]{sf::st_intersection()}}, \code{\link[sf:st_transform]{sf::st_transform()}}
}
