% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_sf.R
\name{clip_sf}
\alias{clip_sf}
\title{Clip feature class to polygon}
\usage{
clip_sf(input_feature, clip_feature)
}
\arguments{
\item{input_feature}{Spatial feature (\code{sf} object) to be clipped.}

\item{clip_feature}{Polygon feature (\code{sf}) used to clip \code{input_feature}.}
}
\value{
An \link[sf:sf]{sf::sf} object
}
\description{
This function clips a \code{sf} object using \code{sf::st_intersection()}. First, this
function checks that the coordinate reference system (CRS) of the input
object is the same as the clipping object. If it is not, this function
transforms the clipping object to CRS of the input object using
\code{sf::st_transform()}before clipping. The output CRS is not changed.
}
\examples{
library("GIStools")

# Read spatial data into R
t_path <- file.path("T:/path/to/project/directory")
gdb_path <- file.path(t_path, "GIS_Data.gdb")
plan_area <- read_fc(lyr_name = "PlanArea", dsn = gdb_path, crs = "NAD83")
roads <- read_fc(lyr_name = "AreaRoads", dsn = gdb_path, crs = "NAD83")

# Clip to extents
plan_area_roads <- clip_fc(roads, roads)
}
\seealso{
\code{\link[mpsgGIStools:read_fc]{mpsgGIStools::read_fc()}}, \code{\link[sf:geos_binary_ops]{sf::st_intersection()}},
\code{\link[sf:st_transform]{sf::st_transform()}}
}
