% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfire_functions.R
\name{get_landfire}
\alias{get_landfire}
\title{Download LANDFIRE Existing Vegetation Type (250EVT)}
\usage{
get_landfire(
  aoi_poly,
  lf_dir = tempdir(),
  email,
  products = "LF2024_EVT",
  res = 30
)
}
\arguments{
\item{aoi_poly}{Area of interest \code{sf} polygon object.}

\item{lf_dir}{Directory path to save raster to. Default is a temporary
directory path using \code{\link[=tempdir]{tempdir()}}.}

\item{email}{Email address. Passed on to \code{\link[rlandfire:landfireAPIv2]{rlandfire::landfireAPIv2()}}.}

\item{products}{Vector of LANDFIRE products. Defult is \code{LF2024_EVT}. See
\url{https://lfps.usgs.gov/products} for available products.}

\item{res}{Raster resolution. Default is 30.}
}
\value{
A list containing 1) the masked raster and 2) the attribute table
for the masked raster.
}
\description{
This function downloads and saves LANDFIRE raster data using the \code{rlandfire}
package and the raster data masked to an area of interest polygon. NOTE:
This function has not been tested for more than one LANDFIRE product.
}
\examples{
\dontrun{
library("psoGIStools")
library("dplyr")

#-- Read spatial data into R
# Dixie National Forest Adminstrative Boundary
dif <- read_edw_lyr("EDW_ForestSystemBoundaries_01") |>
  filter(forestname == "Dixie National Forest")

# Get LANDFIRE EVT data
evt_dif <- get_landfire(dif, lf_dir = file.path("data", "landfire"), 
                        email = "your.name@usda.gov")
}
}
