% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfire_functions.R
\name{get_landfire}
\alias{get_landfire}
\title{Download LANDFIRE Existing Vegetation Type (250EVT)}
\usage{
get_landfire(aoi_polygon, lf_dir, email_address, res = 30)
}
\arguments{
\item{aoi_polygon}{Area of interest \code{sf} polygon object.}

\item{lf_dir}{Directory path to save raster to.}

\item{email_address}{Email address. Passesd on to \code{\link[rlandfire:landfireAPIv2]{rlandfire::landfireAPIv2()}}.}

\item{res}{Raster resolution. Default is 30.}
}
\value{
A list of two data frames summarizing area of EVT.
}
\description{
This function downloads and saves LANDFIRE Existing Vegetation Type (250EVT)
raster data using the \code{rlandfire} package and summarizes the data by EVT
Name and EVT Physiognomy for the area of interest.
}
\examples{
library("GIStools")

# Read spatial data into R
t_path <- file.path("T:/path/to/project/directory")
gdb_path <- file.path(t_path, "GIS_Data.gdb")
plan_area <- read_fc(lyr_name = "PlanArea", dsn = gdb_path)

# Get LANDFIRE EVT data
lf_plan_area <- get_landfire(plan_area, file.path("data", "landfire"), 
                             "your.name@usda.gov")
}
