% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_nwi_data.R
\name{pull_nwi_data}
\alias{pull_nwi_data}
\title{Read FWS National Wetlands Inventory into R}
\usage{
pull_nwi_data(aoi_poly, gdb_path = FALSE, suffix = FALSE)
}
\arguments{
\item{aoi_poly}{\code{sf} polygon object of area of interest.}

\item{gdb_path}{Optional. Character. Directory path to geodatabase. Default
is FALSE. IF FALSE, data will not be saved to a geodatabase.}

\item{suffix}{Optional. Character. A suffix to add to the layer name when
writing to a geodatabase(e.g. "PlanArea").}
}
\value{
A list of \link[sf:sf]{sf} objects: \emph{riparian} and \emph{wetlands}.
}
\description{
\strong{WARNING}. This function is under active development. This function read US
Fish and Wildlife Service National Wetlands Inventory (NWI) data from
the USGS Wetlands ArcGIS REST Service,
https://fwspublicservices.wim.usgs.gov/wetlandsmapservice/rest/, into R
and clips them to an area of interest.
}
\examples{
\dontrun{
library("psoGIStools")
library("dplyr")

# Read in forest boundary
sjf <- read_edw_lyr("EDW_ForestSystemBoundaries_01") |> 
  filter(forestname == "San Juan National Forest")

# Pull NWI data
sjf_nwi <- pull_nwi_data(sjf)
}
}
