% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edw_lyr.R
\name{read_edw_lyr}
\alias{read_edw_lyr}
\title{Read spatial data from Forest Service ArcGIS REST Services}
\usage{
read_edw_lyr(data_name, layer = 0, service = "arcx", crs = "EPSG:4326")
}
\arguments{
\item{data_name}{Character. Name of map layer.}

\item{layer}{Integer. Number of layer to read. Default is  zero (0).}

\item{service}{Character. The public ("arcx") or internal ("arcn") ArcGIS
REST Service code. Default is "arcx". You must be on a Forest Service
network for "arcn" to work.}

\item{crs}{Coordinate reference system (crs). Default is EPSG:4326
(WGS 84).}
}
\value{
An \link[sf:sf]{sf::sf} object or \link[terra:SpatRaster-class]{terra::SpatRaster}.
}
\description{
This function reads spatial features from the Forest Service ArcGIS REST
Services. Either the public ArcGIS REST Service,
https://apps.fs.usda.gov/arcx/rest/services/EDW, or the internal ArcGIS
REST Service, https://apps.fs.usda.gov/arcn/rest/services/EDW, using the
\code{arcgislayers} package. You must be on a Forest Service network to access
data from the interal ArcGIS REST Service.
}
\examples{
\dontrun{
library("psoGIStools")
library("dplyr")

# Administrative Boundary for the Medicine Bow National Forest
mbf <- read_edw_lyr("EDW_ForestSystemBoundaries_01") |> 
  filter(forestname == "Medicine Bow National Forest")

# Read invasive plants data from internal FS REST service 
inv_plant <- read_edw_lyr("EDW_BioInvasivePlant_01", layer = 1, service = "arcn") |> 
  clip_sf(mbf)

# Read basic roads data from public FS REST service
roads <- read_edw_lyr("EDW_RoadBasic_01") |> 
  clip_sf(mbf)
}
}
\seealso{
\code{\link[arcgislayers:arc_read]{arcgislayers::arc_read()}}, \code{\link[sf:st_transform]{sf::st_transform()}}
}
