% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fc.R
\name{read_fc}
\alias{read_fc}
\title{This script contains two functions:
\itemize{
\item read_fc()
\item view_layers()
Read feature class into R.
}}
\usage{
read_fc(lyr_name, dsn, crs = NULL)
}
\arguments{
\item{lyr_name}{Feature class name.}

\item{dsn}{Path to geodatabase that holds \code{lyr_name}.}

\item{crs}{Target coordinate reference system (CRS). Either and
\code{sf::st_crs()} object or accepted input string for \code{sf::st_crs()} (e.g.
"WGS84" or "NAD83"). See \code{\link[sf:st_crs]{sf::st_crs()}} for more details. Default is NULL.
If NULL, resulting \link[sf:sf]{sf::sf} object will not be transformed.}
}
\value{
\code{sf} object
}
\description{
This function uses the \code{sf} package to read a feature class into R from a
geodatabase (*.gdb) using the \code{sf::read_sf()} function. It then checks that
the feature class is in the target coordinate reference system (CRS) and
will transform the feature to the target CRS if it is not.
}
\examples{
\dontrun{
library("psoGIStools")

dat_sf <- read_fc(lyr_name = "feature_name",
                  dsn = file.path("T:/path/to/geodatabase"),
                  crs = "NAD83")
}
}
\seealso{
\code{\link[sf:st_read]{sf::read_sf()}}, \code{\link[sf:st_transform]{sf::st_transform()}}
}
